+++ src/physics/object.cpp
@@ -5,7 +5,7 @@
 #include <iostream>
 
 // Constructor
-Object::Object(glm::vec3 position, float mass, float radius, glm::vec3 velocity)
+Object::Object(glm::vec3 position, double mass, double radius, glm::vec3 velocity)
 {
   this->position = position;
   this->mass = mass;
@@ -15,12 +15,12 @@ Object::Object(glm::vec3 position, float mass, float radius, glm::vec3 velocity)
 }
 
 // Public functions
-void Object::accelerate(const glm::vec3 &acc)
+void Object::accelerate(const glm::dvec3 &acc)
 {
   this->acceleration += acc;
 };
 
-void Object::move(float dt)
+void Object::move(double dt)
 {
   this->velocity += this->acceleration * dt;
   this->position += velocity * dt;
@@ -29,33 +29,33 @@ void Object::move(float dt)
   this->acceleration = glm::vec3(0.f);
 }
 // Getters
-glm::vec3 Object::getPosition() const
+glm::dvec3 Object::getPosition() const
 {
   return this->position;
 }
-glm::vec3 &Object::getPositionRef()
+glm::dvec3 &Object::getPositionRef()
 {
   return this->position;
 }
-glm::vec3 Object::getVelocity() const
+glm::dvec3 Object::getVelocity() const
 {
   return this->velocity;
 }
-float Object::getMass() const
+double Object::getMass() const
 {
   return this->mass;
 }
-float Object::getRadius() const
+double Object::getRadius() const
 {
   return this->radius;
 }
 
 // Setters
-void Object::setVelocity(const glm::vec3 &velocity)
+void Object::setVelocity(const glm::dvec3 &velocity)
 {
   this->velocity = velocity;
 }
-void Object::setPosition(const glm::vec3 &position)
+void Object::setPosition(const glm::dvec3 &position)
 {
   this->position = position;
 }
@@ -63,35 +63,35 @@ void Object::setPosition(const glm::vec3 &position)
 // Public Functions
 void Object::applyGravitation(const Object &other)
 {
-  glm::vec3 dp = other.position - this->position;
+  glm::dvec3 dp = other.position - this->position;
   float distSq = glm::dot(dp, dp);
   if (distSq < EPS)
     return; // Avoid singularity
   float dist = sqrt(distSq);
   float force = G * this->mass * other.mass / distSq;
-  glm::vec3 acc = dp * (force / (dist * this->mass));
+  glm::dvec3 acc = dp * (force / (dist * this->mass));
 
   accelerate(acc);
 }
 
 int Object::handleCollisions(Object &object, Object &object2)
 {
-  glm::vec3 dp = object2.getPosition() - object.getPosition();
-  float dist = sqrt(pow(dp.x, 2) + pow(dp.y, 2) + pow(dp.z, 2));
+  glm::dvec3 dp = object2.getPosition() - object.getPosition();
+  double dist = sqrt(pow(dp.x, 2) + pow(dp.y, 2) + pow(dp.z, 2));
   if (dist < EPS)
     return 0;                                          // Avoid singularity
   if (dist < object.getRadius() + object2.getRadius()) // Simple collision response
   {
-    glm::vec3 rv = object2.getVelocity() - object.getVelocity();
-    glm::vec3 np = dp / dist;
-    float velAlongNormal = glm::dot(rv, np);
+    glm::dvec3 rv = object2.getVelocity() - object.getVelocity();
+    glm::dvec3 np = dp / dist;
+    double velAlongNormal = glm::dot(rv, np);
     if (velAlongNormal > 0)
       return 0;    // They are moving apart
-    float e = 0.8; // Coefficient of restitution
-    float j = -(1 + e) * velAlongNormal / (1 / object.getMass() + 1 / object2.getMass());
+    double e = 0.8; // Coefficient of restitution
+    double j = -(1 + e) * velAlongNormal / (1 / object.getMass() + 1 / object2.getMass());
 
-    glm::vec3 penetration = np * ((object.getRadius() + object2.getRadius() - dist) / 2);
-    glm::vec3 impulse = j * np;
+    glm::dvec3 penetration = np * ((object.getRadius() + object2.getRadius() - dist) / 2);
+    glm::dvec3 impulse = j * np;
 
     object.setVelocity(object.getVelocity() - (impulse / object.getMass()));
     object.setPosition(object.getPosition() - penetration);
